Tweet Retriever, Pig Latin Translator(IP).

+++ src/Calculator/Calculator.java
@@ -1,23 +1,23 @@
 package Calculator;
 
 public class Calculator {
-	public int add(int number1, int number2) {
-		int result = number1 + number2;
+	public double add(double number1, double number2) {
+		double result = number1 + number2;
 		return result;
 	}
 
-	public int subtract(int number1, int number2) {
-		int result = number1 - number2;
+	public double subtract(double number1, double number2) {
+		double result = number1 - number2;
 		return result;
 	}
 
-	public int multiply(int number1, int number2) {
-		int result = number1 * number2;
+	public double multiply(double number1, double number2) {
+		double result = number1 * number2;
 		return result;
 	}
 
-	public int divide(int number1, int number2) {
-		int result = number1 / number2;
+	public double divide(double number1, double number2) {
+		double result = number1 / number2;
 		return result;
 	}
 }
